Refactor: move SysInfo and PerformanceCounter into System module
