+++ src/shared/System/PerformanceCounter.cpp
